html/semantics/forms/the-input-element/image02.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>input type file</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#file-upload-state-(type=file)">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<input type=image id=image>
<script>
  var image = document.getElementById('image'),
      t1 = async_test("setting src to an invalid image fires the error event"),
      t2 = async_test("setting src to an valid image fires the load event");
  
  image.onerror = t1.step_func_done(function(e){
    assert_true(e.isTrusted, "error event is trusted");
    assert_false(e.bubbles, "error event doesn't bubble");
    assert_false(e.cancelable, "error event is not cancellable");
    image.src = "/media/poster.png"; // valid image
  });

  image.onload = t2.step_func_done(function(e){
    assert_true(e.isTrusted, "error event is trusted");
    assert_false(e.bubbles, "error event doesn't bubble");
    assert_false(e.cancelable, "error event is not cancellable");
  });

  image.src = "/media/foobarbarfoo.jpg"; // doesn't exist
</script>
